docker/gemma/requirements.txt
libclang==18.1.1
torch==2.3.0
transformers==4.41.0
datasets==2.19.1
peft==0.11.1
accelerate==0.30.1